refactor: modernize diagnostics and fix clang-tidy findings across codebase
diagnostics:
- replace snprintf with std::format_to_n in minidump (eliminates hicpp-vararg)
- cast MINIDUMP_TYPE operands to unsigned (eliminates hicpp-signed-bitwise)
- fix build_dump_path pass-by-value bug (array was copied, output discarded)
- fix string_view constructed from std::array instead of .data() (wrong length)
- remove orphaned pragma diagnostic pop in stack_walker
- static constexpr → inline constexpr for header constants
- add missing exception codes to exception_code_name (FLT_DENORMAL/INEXACT/STACK_CHECK)
- fix dead store of module_offset in lightweight crash report
- add const-correctness throughout
- de-template build_dump_path (only used with MAX_PATH)
- narrow pragma suppressions to minimal scope
- replace strncpy/memcpy with std::copy_n, memcpy→memcpy with scoped pragma

codebase-wide:
- remove redundant typename (C++20)
- add const-correctness (game_init, parsers, logger, win32)
- suppress -Wcast-function-type-strict in vmp/integrity_bypass
- clang-format all sources

+++ src/diagnostics/crash_handler.cpp
@@ -12,7 +12,7 @@ namespace diagnostics {
         void *g_veh_handle = nullptr;
 
         auto NTAPI veh_handler(EXCEPTION_POINTERS *ep) -> LONG {
-            auto code = ep->ExceptionRecord->ExceptionCode;
+            auto code = static_cast<std::uint32_t>(ep->ExceptionRecord->ExceptionCode);
             if (!is_hardware_exception(code)) {
                 return EXCEPTION_CONTINUE_SEARCH;
             }
